+++ packages/core/src/types/index.ts
@@ -21,8 +21,8 @@ export { DEFAULT_ANIMATION_DURATION } from './route'
 
 export type { NavigationGuardNext, NavigationGuardNextOptions, NavigationRedirectMode, NavigationGuard, PostNavigationGuard } from './guard'
 
-export type { RouterOnError, RouterOptions, Router } from './router'
+export type { RouterOnError, RouterOptions, Router, GuardRouteOptions } from './router'
 
-export type { RouterError, NavigationFailure } from './error'
+export type { RouterError, UniApiCause, UniApiError, NavigationFailure } from './error'
 
 export { RouterErrorCode } from './error'