+++ packages/playground/vite.config.ts
@@ -1,6 +1,6 @@
 import { defineConfig } from 'vite'
 import uni from '@dcloudio/vite-plugin-uni'
-import { generateRouter } from '@meng-xi/vite-plugin'
+import { generateRouter } from '@meng-xi/vite-plugin/plugins/generate-router'
 
 export default defineConfig({
 	plugins: [
@@ -19,7 +19,7 @@ export default defineConfig({
 				navigationBarTitleText: 'title',
 				requireAuth: 'requireAuth'
 			},
-			fileHeader: true,
+			headerTemplate: true,
 			dts: true
 		})
 	]