+++ packages/uni-app/vite.config.ts
@@ -11,7 +11,7 @@ export default defineConfig({
 			pagesJsonPath: 'pages.json',
 			outputPath: 'router.config.ts',
 			dts: 'router.d.d.ts',
-			fileHeader: true,
+			headerTemplate: true,
 			metaMapping: {
 				navigationBarTitleText: 'title',
 				requireAuth: 'requireAuth'